# Conflicts:
#	build-node/3.js
#	build-node/inc.js
#	build/wallet.js
#	build/web/wallet.js
#	lib/core/constants.js
#	privacy.wasm
#	wasm/internal/metadata/bridge/unshield_request.go
#	wasm/internal/metadata/common/constants.go
#	wasm/internal/metadata/exports.go
#	wasm/internal/metadata/parse.go

+++ wasm/internal/metadata/common/utils.go
@@ -436,7 +436,7 @@ type OTADeclaration struct {
 type Uint64Reader uint64
 
 func (u Uint64Reader) MarshalJSON() ([]byte, error) {
-	return json.Marshal(u)
+	return json.Marshal(uint64(u))
 }
 func (u *Uint64Reader) UnmarshalJSON(raw []byte) error {
 	var theNum uint64
